fix(backend): 評審指摘に沿って移行の帯掛け・巻き戻し守衛・NUMBER 設定の int 域検証・合計残高の long 化・過去期限の拒否を正す
- manual_grant_points の drop に t_customers.points と同形の帯掛け（0 以外の行があれば HALT）を追加
- 顧客残高列の巻き戻しに台帳仕訳の存在で止まる守衛を追加（0 復元と台帳削除で残高が静かに失われるため）
- NUMBER 設定の検証を int の範囲まで収斂（読み手は全キー int で読み既定値へ倒すため、保存成功後の意味の入れ替わりを保存時に止める）
- 残高の合計を long 化（明細は int・合計は long。単発付与の上限検査と合わせ int 表現範囲の族を閉じる）
- 正の手動調整で業務日より前の有効期限を拒否（期限当日は可用判定と同じく有効）

+++ backend/src/main/java/com/kizuna/customer/api/dto/CustomerPointBalanceResponse.java
@@ -19,6 +19,6 @@ public class CustomerPointBalanceResponse {
   /** 顧客が会員に紐づいているか。偽なら残高は載らない。 */
   private boolean linked;
 
-  /** 紐づく会員の現在残高。未紐づけでは null。 */
-  private Integer balance;
+  /** 紐づく会員の現在残高。未紐づけでは null。台帳の合計は 1 件の仕訳と違い int を超えうる。 */
+  private Long balance;
 }
